requirements.txt
numpy>=1.24.0
pandas>=2.0.0
aiohttp>=3.9.0
websockets>=12.0
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
tabulate>=0.9.0
colorama>=0.4.6
matplotlib>=3.8.0
scipy>=1.11.0
groq>=0.9.0
